requirements.txt
mistralai
sqlalchemy
google-api-python-client 
google-auth
google-auth-httplib2 
google-auth-oauthlib
streamlit
python-dotenv
